requirements.txt
sagemaker
sagemaker-training
huggingface
flash-attn
transformers
Pillow
numpy
einops
tqdm
dataclasses
timm
matplotlib
torchvision